Move config into a shared directory.
