+++ lsif/src/server/backend/backend.ts
@@ -4,7 +4,7 @@ import { dbFilename, mustGet, NoLSIFDumpError } from '../../util'
 import { XrepoDatabase } from '../../shared/xrepo/xrepo'
 import { TracingContext, logAndTraceCall, addTags, logSpan } from '../../shared/tracing'
 import { Database, sortMonikers, createRemoteUri } from './database'
-import { ConfigurationFetcher } from '../../config'
+import { ConfigurationFetcher } from '../../shared/config/config'
 import {
     DocumentData,
     MonikerData,
